requirements.txt
selenium
channels
Django==3.0.4
django-cors-headers
webdriver-manager
djangorestframework
Pillow
djangorestframework-jwt
coreapi
rest_framework_swagger